アーキテクチャを適用させる
