+++ app/src/main/java/com/hannibal/replacepraeparet/model/Embassy.kt
@@ -1,4 +1,4 @@
-package com.hannibal.replacepraeparet
+package com.hannibal.replacepraeparet.model
 
 data class Embassy(
     var title: String,
