+++ app/src/main/java/com/hannibal/replacepraeparet/view/fragment/DetailFragment.kt
@@ -1,4 +1,4 @@
-package com.hannibal.replacepraeparet
+package com.hannibal.replacepraeparet.view.fragment
 
 import android.graphics.Color
 import android.os.Bundle
@@ -6,8 +6,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.navigation.Navigation
 import com.google.android.material.tabs.TabLayoutMediator
+import com.hannibal.replacepraeparet.model.DetailManager
+import com.hannibal.replacepraeparet.adapter.DetailPagerAdapter
+import com.hannibal.replacepraeparet.model.Flag
 import com.hannibal.replacepraeparet.databinding.FragmentDetailBinding
 
 class DetailFragment : Fragment() {
