+++ app/src/main/java/com/hannibal/replacepraeparet/view/fragment/PostFragment.kt
@@ -1,20 +1,20 @@
-package com.hannibal.replacepraeparet
+package com.hannibal.replacepraeparet.view.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.hannibal.replacepraeparet.databinding.FragmentAccountBinding
+import com.hannibal.replacepraeparet.databinding.FragmentPostBinding
 
-class AccountFragment : Fragment() {
-    private lateinit var binding: FragmentAccountBinding
+class PostFragment : Fragment() {
+    private lateinit var binding: FragmentPostBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        binding = FragmentAccountBinding.inflate(inflater, container, false)
+        binding = FragmentPostBinding.inflate(inflater, container, false)
         return binding.root
     }
 }
