+++ app/src/main/java/com/hannibal/replacepraeparet/view/fragment/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.hannibal.replacepraeparet
+package com.hannibal.replacepraeparet.view.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
