+++ app/src/main/java/com/hannibal/replacepraeparet/viewmodel/DetailEmbassyFragmentViewModel.kt
@@ -1,8 +1,12 @@
-package com.hannibal.replacepraeparet
+package com.hannibal.replacepraeparet.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.hannibal.replacepraeparet.model.Embassy
+import com.hannibal.replacepraeparet.model.Flag
+import com.hannibal.replacepraeparet.model.ScrapingManager
+import com.hannibal.replacepraeparet.model.XmlManager
 import kotlinx.coroutines.launch
 
 class DetailEmbassyFragmentViewModel : ViewModel() {
