+++ app/src/main/res/layout/fragment_information.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InformationFragment"
+    tools:context=".view.fragment.InformationFragment"
     android:id="@+id/container">
 
     <com.google.android.material.tabs.TabLayout
