+++ app/src/main/res/layout/fragment_search.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchFragment">
+    tools:context=".view.fragment.SearchFragment">
 
     <TextView
         android:layout_width="match_parent"
